Extract TransferOrbitFormulaOverlay
